psycopg: introduce new cursor, BSQLQuote intf. for binary protocol
That's the second big step. As discussed 2 years ago with DV,
applications should explicitly choose to go binary, through a
separate interface.

For that reason, introduce the 'cursor_bin' class and BSQLQuote
conformance.

This commit is still incomplete, not all types work so far.

More to come, in some documentation update.