cursor_type: support for C-accelerated dictfetch*() methods
With the supplied test, cursor.dictfetchall() (like in psycopg1) is
proven to be inefficient. The best thing we can do, is to accelerate
the dict-row construction in C, of course.

For that, we allow two things to happen on a normal cursor:
 a. the 'row_factory' can yield a dict type, instead of a sequence.
    Then, this code will populate the dict with column-names and vals.
 b. the 'row_factory' can be itself the dict type. This is a shortcut
    meaning that no constructor function (in python) needs to be called,
    but each row will just be a new dict().

The test in the sandbox proves the performance improvement.